WASD-Movement&CameraFix
Added a script for WASD movement based on the orientation and gave the camera a more stable target. Next : Make a more stable a faster yaw follower, stabilize the camera position, make the hips stay harder when not moving to prevent movement and wiggle while looking around swiftly, enhance the foot placing logic or find a better way.