RagdollFixedRotation
Fixed the issue with the ragdoll understeering, now it nicely follows the capsule. The ragdoll needs visual balancing with a center of mass, hip rotating based on leg position, IK for the rest of the body and a shader or texture so the camera doesnt clip thro the body, but still can see the hands and legs. Also found the problem that was lagging Unity when playing in the editor, it were the gizmos button that needs to be off when playing.