LegBug-FIXED
Finally fixed the bug where the legs would aim in a completely different direction, the targets are placed under the hips cleanly, added new targets to the feet that are just a placeholder because they are parented by the leg target and positioned a bit infront of them. The animated character can now cleanly follow the ragdoll, but the follow speed and accuracy need adjusting, because right now the animated character  isnt positioned perfecly enought, and the ragdoll needs more torque/spring to follow more precisely. After this simple IK is good enough we will move onto adjusting the feet by stepping and walking etc...