BonesPositionMatching
The bones of the ragdoll are now trying to match not only the rotation but also the position of the animated characters bones. The ragdoll is now much more stable and less wigely, will need tweaking with collisions and other stuff but this will help in the IK systems because the error bitween anim and ragdoll will be much smaller, the feets still get stuck because of collisios so we need target correcting (movement up when the error bitween anim legs and ragdoll legs is too big). Btw this features was coded by AI.