NewWorkingLegIK
Just created a new IK script for the legs inspired by the old one, this one works much smoother, added a lot of tweakable values, some of the values are changing based of speed for a smoother look. This script needs the addition of higher speeds such as different curves and settings when running, I think that moving the newPosition target closer to the constantPosition targets when loosing speed or stopping will prevent the legs from trying to reach multiple places at once, this will be crusial for running and other movement.